Point generator tests moved to new project
