+++ cs/TagsCloudVisualization/SpiralPointGenerator.cs
@@ -1,8 +1,9 @@
 ﻿using System.Drawing;
+using TagsCloudVisualization.Interfaces;
 
 namespace TagsCloudVisualization;
 
-public class PointGenerator(Point centerPoint)
+public class SpiralPointGenerator(Point centerPoint) : IPointGenerator
 {
     private const double AngleStep = Math.PI / 360;
     private const double RadiusStep = 0.01;
@@ -10,7 +11,7 @@ public class PointGenerator(Point centerPoint)
     private double _angle;
     private readonly Size _center = new(centerPoint);
 
-    public Point GetNewPoint()
+    public Point GeneratePoint()
     {
         var newX = (int)(_radius * Math.Cos(_angle));
         var newY = (int)(_radius * Math.Sin(_angle));
